move xds_end2end_test to xds subdir, and refactor CountedService
